% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rEAD_fct.R
\name{FID.create.peak}
\alias{FID.create.peak}
\title{create new FID peak}
\usage{
FID.create.peak(bm_inf, bm_sup, pk_mat = gc_ead)
}
\arguments{
\item{bm_inf}{the lower limit}

\item{bm_sup}{the upper limit}

\item{pk_mat}{a mead object}
}
\value{
a mead object
}
\description{
create new FID peak
}
\examples{
# list_ead <- FID.create.peak(pk = 10.756, bm_inf = 10.73, bm_sup = 10.8)
}
