% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rEAD_fct.R
\name{FID.delete.peak}
\alias{FID.delete.peak}
\title{delete FID peak}
\usage{
FID.delete.peak(pk, pk_mat = gc_ead)
}
\arguments{
\item{pk}{RT of peak}

\item{pk_mat}{a mead object}
}
\value{
a mead object
}
\description{
delete FID peak
}
\examples{
# list_ead <- FID.delete.peak(pk = 9.74254)
}
