% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rEAD_fct.R
\name{FID.delete.window}
\alias{FID.delete.window}
\title{delete FID window. Allows you to delete the start or end of a run}
\usage{
FID.delete.window(X1 = "start", X2 = "end", pk_mat = list_ead)
}
\arguments{
\item{X1}{the lower limit or "start" (=T0)}

\item{X2}{the upper limit or "end" (=Tmax)}

\item{pk_mat}{a mead object}
}
\value{
a mead object
}
\description{
delete FID window. Allows you to delete the start or end of a run
}
\examples{
# list_ead <- FID.delete.window(X2 = 5)
# list_ead <- FID.delete.window(X1 = 18)
}
