% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rEAD_fct.R
\name{FID.modify.peak}
\alias{FID.modify.peak}
\title{modify an existing FID peak}
\usage{
FID.modify.peak(pk, bm_inf = NA, bm_sup = NA, pk_mat = list_ead)
}
\arguments{
\item{pk}{time of the peak}

\item{bm_inf}{the new lower limit}

\item{bm_sup}{the new upper limit}

\item{pk_mat}{a mead object}
}
\value{
a mead object
}
\description{
modify an existing FID peak
}
\examples{
# list_ead <- FID.modify.peak(pk = 11.804, bm_inf = 11.715)
# list_ead <- FID.modify.peak(pk = 11.612, bm_inf = 11.526)
# list_ead <- FID.modify.peak(pk = 11.453 , bm_inf = 11.382)
# list_ead <- FID.modify.peak(pk = 11.292, bm_inf = 11.226)
}
