% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rEAD_fct.R
\docType{class}
\name{gcead-class}
\alias{gcead-class}
\title{object used for rEAD package}
\value{
a gcead object
}
\description{
object used for rEAD package
}
\section{Slots}{

\describe{
\item{\code{GC_EAD}}{data.frame.}

\item{\code{peak}}{data.frame.}

\item{\code{pk_res}}{data.frame.}

\item{\code{delay}}{numeric.}

\item{\code{file}}{character.}

\item{\code{num}}{numeric.}

\item{\code{wd}}{character.}

\item{\code{type}}{character.}

\item{\code{param}}{numeric.}
}}

\examples{
#class(gcead)
}
