% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rEAD_fct.R
\name{gcead.graph}
\alias{gcead.graph}
\title{print gcead object}
\usage{
gcead.graph(pk_mat = gc_ead)
}
\arguments{
\item{pk_mat}{a gcead object with spectra and metadata}
}
\value{
a dynamic graph
}
\description{
print gcead object
}
\examples{
# gcead <- gcead.graph()
}
