% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rEAD_fct.R
\name{gcead.merge}
\alias{gcead.merge}
\title{merge a lot of gcead objects in a m(ultiple-gc-)ead}
\usage{
gcead.merge(
  ...,
  gcead_names = TRUE,
  RT_limits = NULL,
  width_smooth = 125,
  wm_bl_fid = 400,
  ws_bl_fid = 10,
  th_pk_fid = 0.01,
  hws_pk_fid = 10,
  snr_pk_fid = 0.5,
  gap_FID = 1
)
}
\arguments{
\item{...}{a suit of gcead objects}

\item{gcead_names}{a vector with new name (optionnal)}

\item{RT_limits}{a couple of decimals for delimit the boundaries of RT (egg : RT_limits = c(4.5,16))}

\item{width_smooth}{width smotthing for "fl" procma::savgol function}

\item{wm_bl_fid}{Width of local window for minimization/maximization for baseline::baseline.rollingBall function}

\item{ws_bl_fid}{Width of local window for smoothing for baseline::baseline.rollingBall function}

\item{th_pk_fid}{threshold of FID peak intensity}

\item{hws_pk_fid}{numeric, half window size for MALDIquant::detectPeaks}

\item{snr_pk_fid}{single numeric value for MALDIquant::detectPeaks. SNR is an abbreviation for signal-to-noise-ratio. A local maximum has to be higher than SNR*noise to be recognize as peak.}

\item{gap_FID}{numeric. Multiplied intensity of FID signal.}
}
\value{

}
\description{
merge a lot of gcead objects in a m(ultiple-gc-)ead
}
\examples{
# list_test <- gcead.merge(b23, b24, b26, b29, b30, b31, gcead_names = c("b23","b24","b26","b29","b30","b31"))
}
