% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rEAD_fct.R
\name{import.GC.EAD}
\alias{import.GC.EAD}
\title{Import data}
\usage{
import.GC.EAD(
  file_csv = "name.csv",
  num = 1,
  wd = NULL,
  delay = 0,
  skip_time = 6,
  FID_calibration = NULL,
  param = c(40, 7, 20, 7),
  skip_pk_cal_GCEAD = NA,
  skip_pk_cal_FID = NA,
  save_fid_cal = TRUE,
  FID_cal = -10^6
)
}
\arguments{
\item{file_csv}{name of the file to be imported}

\item{num}{the spectrum number. Only if several spectra have been concatenated to a single file}

\item{wd}{working directory}

\item{delay}{delays between FID and EAD spectrum}

\item{skip_time}{delete x firsts minutes (X = number of minutes)}

\item{FID_calibration}{calibration file}

\item{param}{parameters of EAD HalfWindowsSize, EAD SNR, FID HalfWindowsSize and FID SNR}

\item{skip_pk_cal_GCEAD}{numeric. Skip peak, higher in first, during the calibration phase.}

\item{skip_pk_cal_FID}{numeric. Skip peak, higher in first, during the calibration phase.}

\item{save_fid_cal}{logical. Save the plot html of calibration transfert from FID to syntech format}

\item{FID_cal}{numeric. Value depend of window width filter for calibration transfert}
}
\value{
a gcead object with spectra and metadata
}
\description{
Import data
}
\examples{
# phase 1
# gcead <- import.GC.EAD(file_csv = "data/sbe18.csv", num = 1, delay = 0, FID_calibration = "data/sbe21_FID2.csv")
# gcead <- print.GCEAD()
}
