% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rEAD_fct.R
\name{m.EAD.average}
\alias{m.EAD.average}
\title{calculate an average for the multi-EAD}
\usage{
m.EAD.average(
  pk_mat = list_ead,
  treatment = c("mean", "median"),
  shift = -1,
  overlay = TRUE
)
}
\arguments{
\item{pk_mat}{a mead object}

\item{treatment}{the type of treatment. Choose "mean" or "median"}

\item{shift}{for shift EAD signals}

\item{overlay}{overlay EAD}
}
\value{
a mead object
}
\description{
calculate an average for the multi-EAD
}
\examples{
# list_ead <- m.EAD.average(list_ead,"mean")
}
