% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rEAD_fct.R
\name{m.EAD.baseline}
\alias{m.EAD.baseline}
\title{delete multi-EAD baseline}
\usage{
m.EAD.baseline(cr_shift = -0.1, pk_mat = list_ead)
}
\arguments{
\item{cr_shift}{for shift the EAD signals}

\item{pk_mat}{a mead object}
}
\value{
a mead object
}
\description{
delete multi-EAD baseline
}
\examples{
# list_ead <- m.EAD.baseline(shift = -0.1)
}
