% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rEAD_fct.R
\name{m.EAD.norm}
\alias{m.EAD.norm}
\title{normalize multi-EAD signal}
\usage{
m.EAD.norm(shift = -1, amplitude = NA, overlay = TRUE, pk_mat = list_ead)
}
\arguments{
\item{shift}{for shift EAD signals}

\item{amplitude}{modulated the amplitude of signals}

\item{overlay}{overlay EAD}

\item{pk_mat}{a mead object}
}
\value{
a mead object
}
\description{
normalize multi-EAD signal
}
\examples{
# list_ead <- m.EAD.norm(shift = -1, amplitude = 2)
}
