% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rEAD_fct.R
\docType{class}
\name{mead-class}
\alias{mead-class}
\title{object used for rEAD package. It's specific for multi EAD.}
\value{
a mead object
}
\description{
object used for rEAD package. It's specific for multi EAD.
}
\section{Slots}{

\describe{
\item{\code{EAD}}{numeric.}

\item{\code{time}}{numeric.}

\item{\code{FID}}{numeric.}

\item{\code{pk_windows}}{numeric.}

\item{\code{bl_EAD}}{numeric.}

\item{\code{cr_EAD}}{numeric.}

\item{\code{peak}}{data.frame.}

\item{\code{delay}}{numeric.}

\item{\code{file}}{character.}

\item{\code{num}}{numeric.}

\item{\code{wd}}{character.}

\item{\code{name}}{character.}

\item{\code{param}}{numeric.}
}}

\examples{
#class(mead)
}
