% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rEAD_fct.R
\name{pk.measure}
\alias{pk.measure}
\title{return information on peaks, amplitude, position, time of depolarisation, SNR}
\usage{
pk.measure(pk_mat = list_ead)
}
\arguments{
\item{pk_mat}{}
}
\value{
a matrix
}
\description{
return information on peaks, amplitude, position, time of depolarisation, SNR
}
\examples{
# depol_param <- pk.measure(list_ead)
}
