% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rEAD_fct.R
\name{recalc.delay}
\alias{recalc.delay}
\title{allows to recalculate delay between FID and EAD signal}
\usage{
recalc.delay(pk_EAD, pk_FID, pk_mat = gc_ead)
}
\arguments{
\item{pk_EAD}{the time of a depolarizing EAD peak}

\item{pk_FID}{the time of a trigger FID peak}

\item{pk_mat}{a gcead object}
}
\value{
the delay between FID and EAD signal.
}
\description{
allows to recalculate delay between FID and EAD signal
}
\examples{
# recalc.delay(pk_FID = 13.688, pk_EAD = 13.623, dl_init = 0) # 386 # dl_init = delai initial
# 386

# phase 2
# gcead <- import.GC.EAD(file_csv = "data/sbe18.csv", num = 1, delay = 386, FID_calibration = "data/sbe21_FID2.csv")
# gcead <- print.GCEAD()
}
