% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rEAD_fct.R
\name{import.GC.EAD}
\alias{import.GC.EAD}
\title{Import data}
\usage{
import.GC.EAD(
  file_csv = "name.csv",
  num = 1,
  wd = NULL,
  delay = 0,
  skip_time = 6,
  FID_calibration = NULL,
  param = c(40, 7, 20, 7),
  skip_pk_cal_GCEAD = NA,
  skip_pk_cal_FID = NA,
  save_fid_cal = TRUE,
  FID_cal = -10^6
)
}
\arguments{
\item{file_csv}{name of the file to be imported}

\item{num}{the spectrum number. Only if several spectra have been concatenated to a single file}

\item{wd}{working directory}

\item{delay}{delays between FID and EAD spectrum}

\item{skip_time}{delete x firsts minutes (X = number of minutes)}

\item{FID_calibration}{calibration file}

\item{param}{parameters of EAD HalfWindowsSize, EAD SNR, FID HalfWindowsSize and FID SNR}

\item{skip_pk_cal_GCEAD}{numeric. Skip peak, higher in first, during the calibration phase.}

\item{skip_pk_cal_FID}{numeric. Skip peak, higher in first, during the calibration phase.}

\item{save_fid_cal}{logical. Save the plot html of calibration transfert from FID to syntech format}

\item{FID_cal}{numeric. Value depend of window width filter for calibration transfert}
}
\value{
a gcead object with spectra and metadata
}
\description{
La fonction importe le file_csv correspondant a la sorti csv de l'instrument
de syntech. Le signal EAD est en colonne 3 et le signal FID en colonne 4. Si
delay est precise, le signal EAD est decale d'autant de pixel.
}
\details{
Le signal FID du syntech est bien souvent sature pour les pics les plus intenses.
Cela oblige a mettre en place plusieurs etapes pour corriger le probleme.
Premierement, il faut rehausse le signal pour que la ligne de base (la mediane)
soit superieur a zero. Puis il faut ramener les parties tronquees a l'intensite maximum.

Ceci fait, la partie FID est chargee. Il y a quatre colonnes le temps en ms,
en min, le RI et le FID. Si skip_time est precise, le debut du FID est tronque
dans cette partie.

Suite a ça, les cinq pics les plus intenses qui ne sont pas satures sont detectes
sur le signal FID de Syntech (GC-EAD). Les cinq meme pics sont detectes sur le
signal FID d'agilent. Cela permet de passer a la phase de d'etalonnage afin de
synchroniser l'echelle de temps avec le signal EAD de syntech.

Enfin, le signal du FID d'agilent est recalcule pour avoir le meme nombre de
point que le signal EAD de syntech.
Dans une derniere partie, les pics sont recalcule et les donnees misent en forme
pour obtenir un seul objet de classe gcead.
}
\examples{
# phase 1
# gcead <- import.GC.EAD(file_csv = "data/sbe18.csv", num = 1, delay = 0, FID_calibration = "data/sbe21_FID2.csv")
# gcead <- print.GCEAD()
}
