% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rEAD_fct.R
\name{mead.graph}
\alias{mead.graph}
\title{print mead object}
\usage{
mead.graph(pk_mat = list_ead, view_all = TRUE, save_htlm = FALSE)
}
\arguments{
\item{pk_mat}{a mead object with spectra and metadata}

\item{view_all}{logical. Add raw EAD if TRUE}

\item{save_htlm}{logical. save the graphe in 'Figures' folder}
}
\value{
a dynamic graph
}
\description{
print mead object
}
\examples{
# list_ead <- gcead.graph(list_ead)
}
